refactor local_peer_id
