+++ client/finality-grandpa/src/communication/tests.rs
@@ -45,10 +45,6 @@ struct TestNetwork {
 }
 
 impl sc_network_gossip::Network<Block> for TestNetwork {
-	fn local_peer_id(&self) -> PeerId {
-		PeerId::random()
-	}
-
 	fn event_stream(&self) -> Box<dyn futures::Stream<Item = NetworkEvent, Error = ()> + Send> {
 		let (tx, rx) = mpsc::unbounded();
 		let _ = self.sender.unbounded_send(Event::EventStream(tx));
