+++ client/network-gossip/src/lib.rs
@@ -68,9 +68,6 @@ mod state_machine;
 
 /// Abstraction over a network. Using this trait instead of `NetworkService` itself makes it easier for unit tests.
 pub trait Network<B: BlockT> {
-	/// Returns the local node's `PeerId`.
-	fn local_peer_id(&self) -> PeerId;
-
 	/// Returns a stream of events representing what happens on the network.
 	fn event_stream(&self) -> Box<dyn futures01::Stream<Item = Event, Error = ()> + Send>;
 
@@ -99,10 +96,6 @@ pub trait Network<B: BlockT> {
 }
 
 impl<B: BlockT, S: NetworkSpecialization<B>, H: ExHashT> Network<B> for Arc<NetworkService<B, S, H>> {
-	fn local_peer_id(&self) -> PeerId {
-		NetworkService::local_peer_id(self)
-	}
-
 	fn event_stream(&self) -> Box<dyn futures01::Stream<Item = Event, Error = ()> + Send> {
 		Box::new(NetworkService::event_stream(self))
 	}
