+++ client/src/light/backend.rs
@@ -62,7 +62,7 @@ pub struct ImportOperation<Block: BlockT, S, H: Hasher> {
 	finalized_blocks: Vec<BlockId<Block>>,
 	set_head: Option<BlockId<Block>>,
 	storage_update: Option<InMemoryState<H>>,
-	_phantom: ::std::marker::PhantomData<(S)>,
+	_phantom: ::std::marker::PhantomData<S>,
 }
 
 /// Either in-memory genesis state, or locally-unavailable state.
